backend/requirements.txt
django
dotenv
djangorestframework
django-cors-headers
Pillow